For... of loops.
As a remedy to some of the for loop’s shortcomings, the ES6 version of JavaScript introduced the shorter and more concise for...of loop. This article will outline the benefits of using for...of loops and walk through how to use a for...of loop to iterate through arrays and strings.